+++ sentry-rails/spec/sentry/rails_spec.rb
@@ -26,6 +26,11 @@ RSpec.describe Sentry::Rails, type: :request do
       expect(app.middleware.find_index(Sentry::Rails::RescuedExceptionInterceptor)).to eq(index_of_debug_exceptions + 1)
     end
 
+    it "propagates timezone to cron config" do
+      # cron.default_timezone is set to nil by default
+      expect(Sentry.configuration.cron.default_timezone).to eq("Etc/UTC")
+    end
+
     it "inserts a callback to disable background_worker for the runner mode" do
       Sentry.configuration.background_worker_threads = 10